feat: add new Squirrel.Mac bundle installation method behind flag
The 'SquirrelMacEnableDirectContentsWrite' user default in your apps defaults suite
can be used to control this new installation method.  It is designed to remove the
requirement that the updating process have write access to its parent directory.

With this feature enabled the updating process only needs write access to it's own
.app bundle folder, not the owning /Applications folder.  This should allow more
non-admin users to update applications when appropriately granted group based permissions.

E.g. 775 and chown :staff